// src/models/regions/locations/spawn/leadership/types.ts
export type LeadershipTraits =
  | 'afflicted'
  | 'capricious'
  | 'charitable'
  | 'corrupt'
  | 'decadent'
  | 'despotic'
  | 'fading'
  | 'imperious'
  | 'incompetent'
  | 'melancholic'
  | 'nascent'
  | 'neglectful'
  | 'occult'
  | 'optimistic'
  | 'overextended'
  | 'paranoid'
  | 'pious'
  | 'respected'
  | 'schismatic'
  | 'treacherous'
  | 'xenophilic'
  | 'xenophobic'
  | 'zealous'